requirements_knowledge.txt
pdfplumber>=0.10.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
google-generativeai>=0.5.0
tiktoken>=0.6.0
python-frontmatter>=1.1.0
python-dotenv>=1.0.0
